+++ kali-config.sh
@@ -1,8 +1,9 @@
 #!/bin/bash
+set -e
 
-rm -rf ~/.zsh-very-old
-mv ~/.zsh-old ~/.zsh-very-old
-mv ~/.zsh ~/.zsh-old
+[ -f ~/.zsh-very.old ] && rm -rf ~/.zsh-very-old
+[ -f ~/.zsh-old ] && mv ~/.zsh-old ~/.zsh-very-old
+[ -f ~/.zsh ] && mv ~/.zsh ~/.zsh-old
 
 cp -r $(pwd)/home/.zsh-kali ~
 cp -r $(pwd)/home/.vimrc ~
@@ -11,10 +12,6 @@ cp -r $(pwd)/home/.config ~
 cp -r $(pwd)/home/.tmux-server.conf ~/.tmux.conf
 touch ~/.local.zsh
 
-echo "source ~/.zsh-kali/zshrc" >> ~/.zshrc
-
-# Install Vim Stuff
-curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
-    https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
-sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
-       https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
+if [ tail -n 1 ~/.zshrc != "source ~/.zsh-kali/zshrc" ]; then
+    echo "source ~/.zsh-kali/zshrc" >> ~/.zshrc
+fi
